requirements.txt
opencv-python>=4.9
scikit-image>=0.22
numpy>=1.26
pillow>=10
matplotlib>=3.8
tqdm>=4.66
scikit-learn>=1.4  # used only for train/val/test splitting in generate_dataset.py
pyyaml>=6.0
scipy>=1.12
# Optional: easyocr>=1.7  (OCR readability demo)
